production_run/requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
requests>=2.31
pyyaml>=6.0
matplotlib>=3.7
lightgbm>=4.0
